+++ CHANGELOG.md
@@ -37,6 +37,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add pull to refresh
 - Add settings page
 - Add recipes
+- Add possiblity to add a recipe to a list
 
 ### Changed